refactor(project-generator): custom artifactId
This refactors project generator to allow for a custom artifactId in the
generated project. This help with generating Maven modules for each
generated project so that several generated projects can be included in
a multi-module Maven build.

This is required for build image generator which will generate an
integration project for each known connector.