fix(s2i): broad inclusion of dependencies
When two connectors, usually by means of transitive dependencies, depend
on the same artefact of different versions, Maven version resolution[1]
will pick the newest of version.

When we build the S2I image we generate an Integration that contains all
known Connectors and Connector templates, with the intent on gathering
all dependencies.

Because of the version resolution only one of the versions needed will
be included. This poses a problem when a Connector depending on a older
version of the artefact is used without the Connector depending on the
newer version of the artefact. The older version would then be fetched
from a Maven repository.

The S2I image should contain all artefacts needed in order to allow
for offline S2I builds. In this case having the older version missing
jeopardizes that.

This addresses that issue by generating a multi-module project with a
module for every Container/Container template in an atempt to circumvent
cases when two Connectors depend on an artefact of different versions.
This way all version of such artefact should be present as the version
resolution is not needed for a Maven module containing only the single
dependency tree for a single Connector.

As there should not be any need to add individual actions to an
Integration whose project is to be built, with regards to gathering
dependencies -- actions do not hold additional dependencies; that bit of
logic was removed, so that single project (Maven module) is generated
for each Connector/Connector template.

Integratio test was added to launch the S2I build without networking to
make sure dependencies needed during the build are contained in the S2I
image. The test builds a trivial Integration, further refinement of the
test could include launching the S2I build for each of the containers.
In as such, the test caught a missing dependency in the S2I image built
without the changes in the image generator builder.

Ref. https://issues.redhat.com/browse/ENTESB-15064

[1] https://cwiki.apache.org/confluence/display/MAVENOLD/Versioning