fix(s2i): make sure cached artifacts are used
When the image for the integration is build via S2I we need to use the
artifacts that have been cached in the `/tmp/artifacts/m2` in the
Syndesis S2I image.

Previously passing `-Dmaven.repo.local=/tmp/artifacts/m2` in the Build
seemed to have done just that, but for _whatever_ reason, perhaps[1],
this started failing recently. Or we noticed it failing recently.

So if the specifying `maven.repo.local` is no longer helpful, the
solution here is to specify the `/tmp/artifacts/m2` repository within
the `settings.xml` placed in `configuration` directory of the generated
project.

[1] https://lists.apache.org/thread.html/f8790f0dcc7195c841fd2717605c96036c82a01d5a45842ac9bfe2fd%40%3Cusers.maven.apache.org%3E