fix: make sure closable resources are closed
When investigating ENTESB-15548[1] I've found several instances of
streams not being closed. Some of these streams are streams that are
obtained from the PostgreSQL JDBC driver and might lead to connection
leak. I haven't investigated if that's the case, other than what's
already noted on ENTESB-15548. I think this refactoring is safe.

[1] https://issues.redhat.com/browse/ENTESB-15548