refactor: also tag the operator with moving tag
For a reason that is lost on me, and could be no longer relevant, we do
not tag/push the operator image with the moving tag, i.e. with x.y, but
only with the release tag. This removes this peculiarity of the operator
release.

Also refactors tagging/pushing of the syndesis-upgrade image to
consolidate the logic.