refactor: use JUnit Jupiter
Thoroughly refactors tests to depend only on JUnit 5 (Jupiter). To make
this transition a stub with empty JUnit classes was needed
(`junit4-stub` module) to help with compiling against Testcontainers and
Citrus framework classes (amongst others).

Crux of the refactor was changing package names for `Test` annotations
and using builtin JUnit 5 parameterized test support.

Some tests that were disabled and unmaintained were outright removed.
This is the case with the system-test module which we do not run and
haven't maintained in years now.