fix: make sure messages are captured in pollenrich
The default aggregation strategy in `PollEnricher` does not persist the
headers needed by Syndesis to perform output message capture. Seems that
aggregation tries to copy data from the `Exchange` created in the
`PollEnrich` but it ends up just copying everything from that `Exchange`
while discarding any values from the previous `Exchange`. We need the
actions that utilize `PollEnrich` to maintain message capture so that
the data is available for mapping. This should also maintain the
activity tracking.