refactor(build): single Maven execution
For building Maven modules when using `-m` invoke Maven several time,
this is slower but also doesn't allow for easy cross-module
dependencies.

This changes the build so that a single Maven invocation is run and the
selection of modules to build is left to Maven to decide based on the
profile names (`build-*`) passed in.

This makes for a faster build but also keeps the module dependencies
within Maven, rather than scripted in shell scripts.