fix: pass header parameters in API client
We filter Camel message headers that we pass on to 3rd party HTTP APIs,
this is to prevent problematic headers to be passed on. We need to block
headers containing sensitive information, or headers that are not
appropriate for the HTTP request. For this reason we use a custom header
filter strategy that contains a list of allowed headers.

Till now we passed only `Content-Type` and `Accept` headers, and did not
pass the headers specified as parameters of an OpenAPI operation.

This changes so for each operation we gather the headers that are given
as parameters and instruct the custom header filter strategy to allow
passing them on to the 3rd party HTTP API.

Fixes #9397