fix: handling of YAML to JSON conversion
We keep all OpenAPI documents as JSON, so when folk upload an OpenAPI
document in YAML format, we convert it to JSON. This conversion used
Jackson `ObjectMaper` configured to drop _empty_ values. So if a
document with empty array for a value would be uploaded the resulting
JSON would not contain the object field containing that value.

This represents a problem when handling OpenAPI documents as some fields
may hold empty arrays, could be due to the OpenAPI generation quirks.

So for example a YAML document:

```yaml
security:
- ApiKeyAuth: []
```

Would be converted to a JSON document:

```json
{
  "security": [{}]
}
```

Loosing the `ApiKeyAuth` field in the process.

This addresses this by configuring a separate `ObjectMapper` for value
conversion and introducing additional robustness in OpenAPI document
handling.

Fixes #9411