fix(credential): support type:name syntax
With #5669 we introduced `type:name` syntax for authentication type. We
needed to follow up on that in #5858, and yet again here with the
credentials support.

When credential provider is to be selected the `type` is what the
provider is keyed against, we need to filter out the `:name` part so
that the correct credential provider can be looked up.