refactor: introduce concept of AuditHandler
With this refactor code from `Auditing` that generates `AuditRecord`s is
delegated to a set of `AuditHandler`s, this way new `AuditHandler`s can
be added without changing or increasing the complexity of `Auditing`.

An `AuditHandler` generates `AuditRecord`s for specific types. This
makes them reusable across different type hierarchies.