chore(build): tweak the non-release profile
The `non-release` profile, previously known as `non-staging` profile is
a profile that's meant to exclude certain modules form building in a
release build. With changes to the syndesis command and the profile
refactoring the profile was built also for modules that do not need it,
such as `ui-react`.
This instigates an extension that allows us to select profiles based on
an expression, which in this case is that the `release` profile is not
active and that the `build-selection`, i.e. building specific modules,
is not done.