fix(build): make sure pull request code
With `pull_request_target` the target branch is checked out, so an
additional `actions/checkout` is needed to checkout the pull request.
So we have been building the code from the target branch `1.12.x` and
not from the pull request. This should fix that by the addition of a
`actions/checkout` when the event is `pull_request_target` that uses the
git id of the pull request.