refactor(server): cleanup
Most of these are around javadoc, which is now enforced by Error Prone.
Some of it fell through the cracks, like resource handling or exception
handling, and minor things like access modifiers.

I've added a fair bit of suppressions around the use of `java.util.Date`
some of these could be refactored, and some of these are mandaded by a
3rd party API.