fix(operator): correct nesting for auditing
The auditing feature toggle was placed after the Jaeger feature toggle,
which did not only change values within `features` key, but terminated
the `features` key and started the `jaeger` top level key. This way when
Jaeger feature was enabled `auditing.enabled` was placed within `jaeger`
and not within `features. This places the `auditing.enabled` before the
Jaeger feature toggle in the server config template, so now the Jager
feature toggle should not affect placement of the auditing feature
toggle. Tests were added to make sure valid (parsable) YAML files are
generated from the generator and that the auditing feature toggle state
is correctly passed to the template.