refactor(build): set `git.commit.id` in all cases
When building/running a subset of modules without `syndesis-parent`
module `git.commit.id` will not be set and will trigger enforcer
failure. This removes `<inherited>false</inherited>` so that the
`git-commit-id` plugin is run in all modules not just in parent.

Previously we have seen performance issues with running the
`git-commit-id` plugin, and confined the execution just to the parent
module. With 4.0.4 it seems those performance are not present, so we can
run it in all modules.