fix(c/soap): change xsd:timeInstant to xsd:dateTime
SOAP 1.2 renamed `xsd:timeInstant` to `xsd:dateTime` see[1]. Using
`xsd:timeInstant` was breaking the tests with AtlasMap as the XML Schema
data types no longer define `xsd:timeInstant`.

Might be a slight concern with regards to legacy SOAP 1.1 or earlier,
this was an easier fix and it's fairly safe that most deployments are
using SOAP 1.2 these days.

[1] https://www.w3.org/2000/xp/Group/1/10/11/soap12-part1.html#schemachanges