fix(c/soap): NPE in XmlSchemaSerializer
This fixes the `NullPointerException` in `XmlSchemaSerializer` reported
in ENTESB-16464[1]. The namespace context in the target schema was not
initialized and defaulted to `null`, leading to the
`NullPointerException`.

In addition to setting the namespace context to an instance of
`NamespaceMap`, this instance was populated with prefixes for XML Schema
and XML namespaces, as these were not set by default.

[1] https://issues.redhat.com/browse/ENTESB-16464