fix(c/soap): add definition of XML schema namespace
XML namespace is an odd one, it's implicitly defined in a XML schema
document, but not in non-XML schema documents. So in the schema sets
created by the SOAP API generator and expected by AtlasMap in the data
shape it needs to be declared and included. Otherwise attributes with
type of `xml:*` can't be unreferenced and the inspection of the data
shape fails.

To circumvent this, XML namespace schema is added automatically to all
schema sets. This might seem unwarranted, but considering that most SOAP
APIs will declare a SOAP fault, and that the `reasonText` is defined
with `xml:lang` attribute inclusion seems warranted.