fix(c/rest): header parameter passing
The `i.s.c.rest.swagger.HeaderParametersCustomizer` in cooperation with
`i.s.i.runtime.util.SyndesisHeaderStrategy` filters out any HTTP headers
that are not provided as parameters of an REST API operation. This is to
prevent passing sensitive or protocol unfriendly HTTP headers to 3rd
parties.

`HeaderParametersCustomizer` marks a HTTP header as allowed if from the
the parameter is given as `in=header` in the operation or in the path/
global parameter list of the OpenAPI document for the API.

To minimize the load on the database and the payloads moved between UI/
server, the OpenAPI document is minimized by removing properties that
are not required for the operation of the API client. In that
minimization HTTP headers were erroneously removed. This had the end
result of not allowing any HTTP headers specified by the API, and
subsequently not sending them in the HTTP payload.

This changes the minimization of OpenAPI documents to make sure HTTP
header parameters are included and adds a end-to-end integration test
that generates an custom REST API client from the OpenAPI document. This
is to prevent future regressions either in the REST API client connector
generation or in the parameter passing to the REST API connector
implementation.

Fixes #9638