chore: data shape parameters
The CSV support in AtlasMap allows specifying additional parameters that
influence how the data shape is handled. One such example is whether
header row is provided in the shape definition, or a parameter that
defines names of columns (headers).

For this the `DataShape` object needed additional property: `parameters`
which holds these additional parameters and passes it to the, in this
case, CSV service to generate the inspected document definition used by
the mapper to present mappings in the UI.

When user has the ability to set the data shape, i.e. for any data shape
of kind `ANY`, on the screen to specify the data shape for the CSV data
shape, additional button `Parameters` is presented opening the
`ParametersDialog` from AtlasMap.