chore(release): moving tag for snapshot release
Instead of stomping the release's moving tag this makes the snapshot
releases (daily releases) have the moving tag with the `-prerelease`
suffix.

Say we just released a version 1.2.3, with it a moving tag 1.2 will be
created. The idea of the moving tag is that we can create minor releases
and have environments auto-update to them. With snapshot (daily)
releases we used to use the same moving tag, so a snapshot release would
stomp over the 'proper' release's moving tag. This prevents that by
using a suffix on the snapshot release's moving tag.