chore(release): create or update GitHub release
Snapshot (daily) releases auto-create a GitHub release, whereas a
'proper' release is initiated by creating a GitHub release. So instead
of failing on the GitHub release creation, as one already exists, this
performs a create-or-update of a GitHub release. This way we can have
snapshot releases fully automated and 'proper' releases triggered via
GitHub UI.