chore(ui): upgrade prettier and add JSON sort
I wanted to update the openapi.internal.json with a trivial change I
made in the server API. The diff for that was too great to be of any
value, so I found a way to add JSON sorting to prettier. Now the JSON
files are sorted so when changes are made only the lines affected by the
change are in the diff.

I've also formatted `openapi.*.json` files, to build the change on
later.