chore(build): enforce Java 1.8 version
Maven compiler plugin hides certain `javac` error messages, for instance
those that are not strictly related to issues compiling Java code. This
makes it very difficult to debug issues. A common cause for `javac` to
fail is using a JVM newer than 1.8. This won't work due to how we use
and configure Errorprone. The configuration is JVM 1.8 specific.