fix(build): make sure `unpack` is not skipped
Using the `dependency-maven-plugin` we unpack an additional prettier
plugin to sort JSON files while formatting. The execution of the
`unpack` goal would be skipped when building with the `flash` profile,
and we need it to run in all cases for the formatting using prettier to
work. So this sets `<skip>false</skip>` in the plugin's configuration.