chore(operator): rework the database upgrade
This reworks the database upgrade not to rely on the existence of
`POSTGRESQL_UPGRADE=copy` functionality. Now the operator performs the
process of upgrading the database by spinning up the new database
instance and loading a dump of the data created from the old database
instance in order to create the data files in the format of the new
version to replace the data files of the old instance.

Several small refactorings were performed, namely utilizing the typed
kubernetes go-client more.

Ref. https://issues.redhat.com/browse/ENTESB-16734