fix: maintain data shape parameters
Even though an issue in AtlasMap was resolved[1] it doesn't seem to help
in the Syndesis use case. This refactors the support for providing data
shape parameters and uses one dirty trick to avoid resetting the
internal state of AtlasMap's `ParametersDialog`. The key is in setting
the `enabled=true` parameter which should leave the value present in the
dialog. The refactoring focused on removing the complexity of AtlasMap
to the AtlasMap adapter so that pages that use the
`DataShapeParametersDialog` can have a simpler interface and need not
take any concern of the AtlasMap internal state.

Ref. https://issues.redhat.com/browse/ENTESB-17416

[1] https://github.com/atlasmap/atlasmap/issues/2990