refactor: support more than one action summary
Reducing action summary to a single summary doesn't suit SOAP client
connectors, as within a WSDL file there could be many services with
different operations, a contrast to OpenAPI documents where there could
be only one API service per document.

So instead of summing up all operations from all services we can now
have one action summary per service.

So when importing a WSDL file with multiple services we can now show the
number of actions along with the number of services. Further refactoring
might opt to show a number of actions per service along with the name of
the service, but lets consider this a first step.

And this is all so we can show the correct number of actions/services
and set the default serviceName/portName/address when the WSDL file
contains more than one service.