fix: create connector with the selected address
We did not persist the selected address when creating the SOAP client
connector, we reverted to the default address set to the first service
in the WSDL file.

This makes sure that the address is passed in `configuredProperties`
when creating the connector.

For this an additional property was needed that contains a map of port
to address so that the correct address for the selected service/port
combination is presented on the last (details) screen of the wizard.