fix: don't throw an error if errors are returned
Throwing an exception here breaks the wizard and reverts to the upload-
-provide-url screen. This is especially awkward when user introduces an
error in the OpenAPI document by editing. In some (most) cases if the
OpenAPI document can be parsed by Apicurio Studio the error can be
corrected and user doesn't need to start from the beginning.

Most notably changing the `openapi` key makes the document unparsable
and it leads to loss of data. Some changes, like changing the `type` in
the object's JSON schema are recoverable.

We should be throwing `Error`s only in exceptional circumstances and
errors reported from the server are not that.

This changes the flow, so the notification of an error will not be shown
but rather the user will land on the first page of the wizard, which
allows the user to edit the OpenAPI document and resolve the issue.