chore: remove `demoData` flag from Syndesis CR
This removes the `demoData` flag from Syndesis Custom resource. The same
effect can now be achieved by enabling the `todo` addon.

Since the operator merge code doesn't support map entry removal, only
value replacement and loosely entry addition. Properties relating to the
sampledb database on resources, namely `DeploymentConfig`s for
`syndesis-server` and `syndesis-db` need to remain. Their values are
changed so that the functionality of enabling/disabling the `todo` addon
is performed.

To apply the change to the PostgresSQL service, post-start lifecycle
hook in `syndesis-db` `DeploymentConfig` is changed to trigger the
redeployment. Otherwise changing the `syndesis-sampledb-config`
`ConfigMap` would not take effect. This causes, hopefully a brief,
downtime of the PostgreSQL pod, which could be avoided if the operator
did not apply changes via the templating system, and it would connect to
the database to apply the changes. This is left as an exercise for
later to make this change as simple as possible.

Ref. https://issues.redhat.com/browse/ENTESB-15863