chore: optimize custom connector specification ...
...persistence and processing

When passing specification resource received in `POST /connector/custom`
to the connector generator, pass the stream instead of String. For large
specifications string concatenation uses large amounts of memory. The
connector generator now needs to read the stream of the specification
from `ConnectorSettings` or if it's not present there can examine
`configuredProperties` to see if the UI provided the specification
there.

When storing the connector, when the connector generator has opted not
to store the specification in the `specification` property of
`configuredProperties` and a specification was provided on the above
endpoint -- signals that the specification should be stored in the file
store (blob) table instead. For this a `SpecificationResourceDao` is
used to manage persistence of the specification and the `specification`
property set in the `configuredProperties` is set to the value of
`db:<connectorId>`.