chore: make sure to pass the specification in POST
The `POST /connector/custom` needs to contain the `specification`
mime-part with the specification provided by the end user if the
specification was not returned in the base connector created on the
`POST /connectors/custom/info` endpoint used first in the custom
connector wizard.

Not including the specification in `configuredProperties` indicates that
the connector generator needs the specification to be repeated in the
`POST /connector/custom` to generate the connector.