chore: use specification stream in SOAP generator
In the custom SOAP client connector generator we can now utilize the
specification resource provided via `ConnectorSettings` instead of the
`specification` as String in `configuredProperties`. This way we utilize
far less memory. Also we don't store the specification in the generated
connector's `configuredProperties` so that the `CustomConnectorHandler`
stores it in the file store (blob) table instead.