chore: dynamic shapes for custom SOAP connectors
With this the data shapes are not embedded in the connector definition
for custom SOAP connectors. The large `specification` property of the
data shape uses a large amount of memory when generated, serialized and
parsed, and also increases the load on the JSONDB.

Instead of embedding them, they're computed dynamically using the
dynamic metadata support, which was extended to support locally run
metadata actions. Without this support any dynamic metadata request
would be forwarded to the meta service. Now local metadata action will
be used for any connector ID that resembles a ID generated by the key
generator -- a common trait for generated connectors, and
`ConnectorAndActionGenerator` will be looked up for request delegation
in application context.

The `ConnectorGenerator` contract was extended via
`ConnectorAndActionGenerator` to allow generating the connector actions
on demand.

The SOAP model generator/parser's action generation was extracted and
exposed for the `SoapApiConnectorGenerator` to use for generating
action, including the data shapes, on demand.