chore: skip `specification` property from database
Seems that there is no need to pass `specification` nor `wsdlURL` to the
CXF component, those properties seem to be optional when in `PAYLOAD`
data format mode. None the less there could be a good reason(TM) for
requiring them, that hasn't dawned on me.

In light of this, this doesn't write the WSDL file from the
`specification` property to the temporary file if that specification is
pointing to the database. In essence proving that we don't require
either `wsdlURL` or `specification` set.