fix: configured property priority
Properties that the `ConnectorStepHandler` generates, that are to be set
on the Camel endpoints, don't reflect the hierarchy of the integration
model. For example configured properties of a Connector have higher
priority than configured properties of a Connection. This makes it
impossible for the end users to have multiple Connections for the same
Connector with different properties -- which is mostly impacting custom
Connectors, especially SOAP connector which allows, for whatever reason,
configuring of credentials (e.g. username/password) on the Connector in
addition to Connection.