refactor: integration scheduling from ConfigMap
Seems odd to tie the server back to the Syndesis CR, it adds an
additional dependency that we can do without and increases the
complexity. This removes the Syndesis CRD definition from the server and
uses the values from the ConfigMap instead. The net effect of this
should not be visible in any behavioral change, e.g. ConfigMap is
watched and the configuration will reload if Syndesis CR is changed and
the operator updates it.