chore: support lowercase table names with Oracle
When dealing with Oracle database metadata it seems that we need to
provide the table/column names in upper case. This adds an additional
variant, to the one variant we already use with PostgreSQL which seems
to need lower case names, to uppercase the names.

The integration test with Oracle runs over 5min due to a very long
database startup and since the Oracle XE image is not available for pull
from an official source, the test is enabled only if the image is built
locally from the Dockerfiles available at[1].

Ref. https://issues.redhat.com/browse/ENTESB-17710

[1] https://github.com/oracle/docker-images/tree/main/OracleDatabase/SingleInstance