fix: keep track of removed environment variables
We don't know if an environment variable in the integration's
DeploymentConfig was removed in the UI or if it was added by manually by
editing the DeploymentConfig as we keep track only of the next wanted
state of the environment variables. So when an environment variable is
removed via the UI we can't detect that and remove it from the
DeploymentConfig.

This keeps track of removed environment variables on integration update
by examining the current state and the next state of the integration and
looking at the difference in environment variables. So when an
environment variable is removed using the UI we now know that we also
need to remove it from the DeploymentConfig.

Ref. https://issues.redhat.com/browse/ENTESB-17748