chore: upgrade operator-sdk to v1.14.0
This seems to trickle down and upgrade quite a bit of dependencies as
well. We can't upgrade logr to v1.2.0 (from 0.4.0) as it seems to have a
breaking change in it as `sigs.k8s.io/controller-runtime/pkg/log` and
`k8s.io/klog/v2` would need to upgrade to it as well.

It builds and tests pass, not sure if an additional step is needed to
regenerate the scaffolded code.