fix: manage environment variables on publish
Instead of managing the environment variables, i.e. in this case removal
of environment variables, at the point the integration is updated this
manages the environment variables at the point of publishing by
examining environment variables of previous deployments and adding any
environment variables that are not present on the latest deployment to
the list of environment variables to remove.

Ref. https://issues.redhat.com/browse/ENTESB-17748