refactor: move from FMP to KOMP
This replaces the Fabric8 Maven Plugin with Kubernetes/Openshift Maven
Plugin.

We used to use `fabric8.mode` to determine which mode of the FMP plugin
will be used: kubernetes - where a Docker daemon builds images;
openshift - when we use the OpenShift builds to build images and push
them to image streams (used in development). The KOMP doesn't have a
mode setting, the mode is implicitly determined by which plugin is used
i.e. `kubernetes-maven-plugin` or `openshift-maven-plugin`, for this
reason we no longer have one `image` Maven profile, but two, aptly named
`image:kubernetes` and `image:openshift`. Both activated by setting
`build.mode` property to either `kubernetes` or `openshift`. This is
handled automatically by the `syndesis` CLI.

In the UI mode, an issue was hit in using OpenShift build mode with a
Dockerfile and an inline assembly:
```
Failed to execute the build: Unable to build the image using the
OpenShift build service: This image has already been flattened, you can
only flatten the image once
```
For this reason the inline assembly was replaced with a symbolic link
from the Docker context directory (`app/ui-react/docker`) to the build
artifacts directory `app/ui-react/syndesis/build`. A better solution
should be found for this.

Since the TAR file generated by the plugin serving as the context sent
for the build is now prefixed with `maven`, `Dockerfile`s were updated
to reflect that.

Ref. https://issues.redhat.com/browse/ENTESB-17787